media_profiles: Remove invalid resolution for FFC
Our front facing camera only supports up to VGA video quality (640x480). In KitKat, the new camera app is setting the default resolution for the FFC in camcorder mode to 480P, which subsequently causes the camera app to force close. This is one of the factors contributing to the force close when switching to FFC in camcorder mode.

Signed-off-by: Sultanxda <sultanxda@gmail.com>